emotion-recognition-system/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
pillow==10.1.0
